+++ src/cpp/xmlparser/XMLProfileManager.cpp
@@ -334,7 +334,7 @@ XMLP_ret XMLProfileManager::extractProfiles(up_base_node_t profiles, const std::
         }
     }
 
-    profileCount += m_transport_profiles.size(); // Count transport profiles
+    profileCount += static_cast<unsigned int>(m_transport_profiles.size()); // Count transport profiles
 
     if (0 == profileCount)
     {
